fix(uffs-cli): bound ps calls in --status so it can't hang
--status shelled out to `ps` twice (process list + parent-name lookup)
via Command::output() with no timeout, so a process with a huge argument
vector (or one stuck in uninterruptible I/O) made `ps` block and hung the
whole command. Route both through a shared capture_with_timeout helper
(2s watchdog on a helper thread; kills the child and reports "process
scan timed out — skipped" on timeout). Require the real `--mcp` flag when
matching sessions so build tooling (cargo/sccache compiling uffs-mcp) is
not listed. Split the MCP-stdio scan into a process_scan submodule to
stay under the file-size gate.
